fix buf

+++ Part1/GUI/GUI_InputForm.cs
@@ -15,18 +15,6 @@ namespace Part1
         {
             InitializeComponent();
 
-            // if user login as Admin, then show the Admin control tab
-            var acc = GetCurrentAccLoginDetail();
-            this.tabControl1.Controls.Remove(this.tabPage4);
-            if (acc.Role.Trim().ToLower() == "admin")
-            {
-                this.tabControl1.TabPages.Add(this.tabPage4);
-                DeleteAccBtn.Location = new Point(-1000, -1000);
-            } else
-            {
-                DeleteAccBtn.Location = new Point(11, 341);
-            }
-
             // set up and hide some buttons
             this.StartPosition = FormStartPosition.CenterScreen;
             this.comboBox1.SelectedIndex = 1;
@@ -35,6 +23,12 @@ namespace Part1
             DeleteBTN.Location = new Point(-1000, -1000);
             RcvBTN.Location = new Point(-1000, -1000);
             DltBTN.Location = new Point(-1000, -1000);
+            AdminDeleteUserBTN.Location = new Point(-1000, -1000);
+
+            // set up Homepage
+            List<Button> buttons = new List<Button> { encodeBTN, IncreaseBTN, DecreaseBTN, DeleteBTN, ResetBTN };
+            StringService.InitializeToolTips(buttons);
+            StringService.SetupDataGridView(dataGridView);
 
             // set up User account info to textboxes in Settings tab
             List<TextBox> textBoxesInput = new List<TextBox>
@@ -45,21 +39,13 @@ namespace Part1
                     DateIn4,
                 };
             AccountService.BindLoggedInAccountToTextBoxes(textBoxesInput);
-
-            // set up Homepage
-            List<Button> buttons = new List<Button> { encodeBTN, IncreaseBTN, DecreaseBTN, DeleteBTN, ResetBTN };
-            StringService.InitializeToolTips(buttons);
-            StringService.SetupDataGridView(dataGridView);
-            StringService.LoadAllDataFromDB(dataGridView);
-            StringService.CountRecords(toolStripStatusLabel1);
+            AccountService.UpdateAccIDTitle(groupBox6);
 
             // set up Recycle Bin
             RecycleBinService.SetupDataGridView(BinGridView);
-            RecycleBinService.LoadAllDataFromDB(BinGridView);
 
             // set up Admin
             AccountService.SetupDataGridView(AccGridView);
-            AccountService.LoadDataToGridView(AccGridView);
 
             // Load Previous Settings
             SettingService.LoadSettings();
@@ -71,6 +57,44 @@ namespace Part1
             SettingService.BindSettingToControls(RBdarkmodeON, RBdarkmodeOff, RBWindow, RBFullscreen, Width, Height, Opacity);
             SettingService.LoadResolution(res.Width, res.Height, this);
             SettingService.CenterTabControl(this);
+        }
+
+        protected override void OnLoad(EventArgs e)
+        {
+            base.OnLoad(e);
+
+            var acc = GetCurrentAccLoginDetail();
+            this.tabControl1.Controls.Remove(this.tabPage4);
+            if (acc.Role.Trim().ToLower() == "admin")
+            {
+                this.tabControl1.TabPages.Add(this.tabPage4);
+                DeleteAccBtn.Location = new Point(-1000, -1000);
+            }
+            else
+            {
+                DeleteAccBtn.Location = new Point(11, 116);
+            }
+        }
+
+        protected override void OnShown(EventArgs e)
+        {
+            base.OnShown(e);
+
+            // Load Homepage DataGridView
+            StringService.LoadAllDataFromDB(dataGridView);
+            StringService.CountRecords(toolStripStatusLabel1);
+
+            // Load Recycle Bin DataGridView
+            RecycleBinService.LoadAllDataFromDB(BinGridView);
+            RecycleBinService.CountRecords(toolStripStatusLabel2);
+
+            // Load Admin DataGridView (only for Admin)
+            var acc = GetCurrentAccLoginDetail();
+            if (acc.Role.Trim().ToLower() == "admin")
+            {
+                AccountService.LoadDataToGridView(AccGridView);
+                AccountService.CountRecords(toolStripStatusLabel3);
+            }
         }
 
         private Account GetCurrentAccLoginDetail() 
@@ -127,6 +151,12 @@ namespace Part1
         {
             List<Button> buttons = new List<Button> { RcvBTN, DltBTN };
             RecycleBinService.CheckAndUpdateRecycleBinButtons(BinGridView, buttons);
+        }
+
+        private void ShowAdminToolButtons()
+        {
+            List<Button> buttons = new List<Button> { AdminDeleteUserBTN };
+            AccountService.CheckAndUpdateAdminButtons(AccGridView, buttons);
         }
 
         private bool isBinding = false;
@@ -246,8 +276,6 @@ namespace Part1
                 StringService.CheckAndUpdateDeleteButton(dataGridView, DeleteBTN);
                 isBinding = false; // end binding
             }
-
-
         }
 
         private void Search_TextChanged(object sender, EventArgs e)
@@ -273,8 +301,7 @@ namespace Part1
             StringService.CountRecords(toolStripStatusLabel1);
             StringService.ClearTextBoxes(textBoxS, textBoxN);
             RecycleBinService.LoadAllDataFromDB(BinGridView);
-
-
+            RecycleBinService.CountRecords(toolStripStatusLabel2);
         }
 
         private void radioButton10_click(object sender, EventArgs e)
@@ -336,11 +363,14 @@ namespace Part1
                 if (AccGridView.SelectedRows.Count > 0)
                 {
                     AccBindingData();
+                    AccountService.UpdateGroupTitle(groupBox8, AccGridView);
                 }
                 else
                 {
-                    ClearBinTextBoxes();
+                    ClearAccTextBoxes();
+                    AccountService.ClearGroupTitle(groupBox8);
                 }
+                ShowAdminToolButtons();
                 isBinBinding = false; // end binding
             }
         }
@@ -354,9 +384,12 @@ namespace Part1
             if (!IsConfirmed(originalID, "restore")) return;
 
             RecycleBinService.RestoreFromRecycleBin(BinGridView);
-            ClearBinTextBoxes();
             RecycleBinService.ClearGroupTitle(groupBox5);
+            RecycleBinService.CountRecords(toolStripStatusLabel2);
+
+            StringService.CountRecords(toolStripStatusLabel1);
             StringService.LoadAllDataFromDB(dataGridView);
+            ClearBinTextBoxes();
         }
 
         private void BinDltBTN_Click(object sender, EventArgs e)
@@ -369,6 +402,7 @@ namespace Part1
             RecycleBinService.DeleteFromRecycleBin(BinGridView);
             ClearBinTextBoxes();
             RecycleBinService.ClearGroupTitle(groupBox5);
+            RecycleBinService.CountRecords(toolStripStatusLabel2);
             ShowBinToolButtons();
         }
 
@@ -387,6 +421,9 @@ namespace Part1
             ClearBinTextBoxes();
             RecycleBinService.ClearGroupTitle(groupBox5);
             StringService.LoadAllDataFromDB(dataGridView);
+
+            StringService.CountRecords(toolStripStatusLabel1);
+            RecycleBinService.CountRecords(toolStripStatusLabel2);
         }
 
         private void BinSearchingTB_TextChanged(object sender, EventArgs e)
@@ -493,6 +530,9 @@ namespace Part1
         {
             SettingService.ResetToDefault(this, new TabPage[] { tabPage1, tabPage2 },
                                RBdarkmodeOff,RBWindow, Width, Height, Opacity);
+            StringService.LoadAllDataFromDB(dataGridView);
+            RecycleBinService.LoadAllDataFromDB(BinGridView);
+            AccountService.LoadDataToGridView(AccGridView);
         }
 
         private void UpdateAccInfo_Click(object sender, EventArgs e)
@@ -515,6 +555,9 @@ namespace Part1
                 int ID = (int)AccGridView.SelectedRows[0].Cells["ID"].Value;
                 AccountService.DeleteUserAccByAdmin(ID);
                 AccountService.LoadDataToGridView(AccGridView);
+                AccountService.CountRecords(toolStripStatusLabel3);
+                ClearAccTextBoxes();
+                AccountService.ClearGroupTitle(groupBox8);
             }
         }
 
@@ -523,8 +566,15 @@ namespace Part1
             var confirm = MessageBox.Show("Are you sure to delete all user account, can't recover back!", "Confirm", MessageBoxButtons.YesNo);
             if (confirm == DialogResult.Yes)
             {
+                // if null
+                if (AccGridView.Rows.Count == 0)
+                {
+                    MessageBox.Show("No Account to move!", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    return;
+                }
                 AccountService.DeleteAllUserAcc();
                 AccountService.LoadDataToGridView(AccGridView);
+                AccountService.CountRecords(toolStripStatusLabel3);
             }
         }
     }
